Refactor files and folders of the project
