+++ app/page.tsx
@@ -1,10 +1,10 @@
-import HeroSection from '@/app/components/home/HeroSection';
-import ServicesGrid from '@/app/components/home/ServicesGrid';
-import HowItWorks from '@/app/components/home/HowItWorks';
-import RoomMakeovers from '@/app/components/home/RoomMakeovers';
-import AppPromotion from '@/app/components/home/AppPromotion';
-import ServicePackages from '@/app/components/home/ServicePackages';
-import Footer from '@/app/components/layout/Footer';
+import HeroSection from '@/components/home/HeroSection';
+import ServicesGrid from '@/components/home/ServicesGrid';
+import HowItWorks from '@/components/home/HowItWorks';
+import RoomMakeovers from '@/components/home/RoomMakeovers';
+import AppPromotion from '@/components/home/AppPromotion';
+import ServicePackages from '@/components/home/ServicePackages';
+import Footer from '@/components/layout/Footer';
 export default function Home() {
     return (
         <main className="min-h-screen pt-20">
