+++ components/home/HowItWorks.tsx
