+++ components/ui/ImageBoxGrid.tsx
@@ -34,6 +34,7 @@ export function ImageBoxGrid({
                             src={item.image}
                             alt={item.title}
                             fill
+                            sizes="(max-width: 768px) 100vw, (max-width: 1200px) 50vw, 33vw"
                             className="object-cover transition-transform group-hover:scale-110"
                         />
